requirements.txt
websocket-client
kafka-python-ng
pyspark==3.5.0
streamlit
pandas